Move useDocsConfig into DocsLayout
