+++ app/routes/query.$version.docs.tsx
@@ -9,8 +9,10 @@ import {
   colorTo,
   textColor,
   repo,
-  useQueryDocsConfig,
   latestVersion,
+  availableVersions,
+  frameworks,
+  localMenu,
 } from '~/projects/query'
 import { getTanstackDocsConfig } from '~/utils/config'
 
@@ -18,10 +20,10 @@ export const loader = async (context: LoaderFunctionArgs) => {
   const { version } = context.params
   const branch = getBranch(version)
 
-  const tanstackDocsConfig = await getTanstackDocsConfig(repo, branch)
+  const config = await getTanstackDocsConfig(repo, branch)
 
   return json({
-    tanstackDocsConfig,
+    config,
     version,
   })
 }
@@ -34,8 +36,7 @@ export const meta: MetaFunction = () => {
 }
 
 export default function RouteFrameworkParam() {
-  const { tanstackDocsConfig, version } = useLoaderData<typeof loader>()
-  let config = useQueryDocsConfig(tanstackDocsConfig)
+  const { config, version } = useLoaderData<typeof loader>()
 
   return (
     <>
@@ -47,6 +48,9 @@ export default function RouteFrameworkParam() {
         colorTo={colorTo}
         textColor={textColor}
         config={config}
+        frameworks={frameworks}
+        availableVersions={availableVersions}
+        localMenu={localMenu}
       >
         <Outlet />
       </DocsLayout>
