+++ app/routes/ranger.$version.docs.tsx
@@ -1,12 +1,14 @@
 import { Outlet, json, useLoaderData } from '@remix-run/react'
 import {
-  useRangerDocsConfig,
   repo,
   getBranch,
   colorTo,
   latestVersion,
   colorFrom,
   textColor,
+  availableVersions,
+  localMenu,
+  frameworks,
 } from '~/projects/ranger'
 import { seo } from '~/utils/seo'
 import { DocsLayout } from '~/components/DocsLayout'
@@ -17,10 +19,10 @@ export const loader = async (context: LoaderFunctionArgs) => {
   const { version } = context.params
   const branch = getBranch(version)
 
-  const tanstackDocsConfig = await getTanstackDocsConfig(repo, branch)
+  const config = await getTanstackDocsConfig(repo, branch)
 
   return json({
-    tanstackDocsConfig,
+    config,
     version,
   })
 }
@@ -33,8 +35,8 @@ export const meta: MetaFunction = () => {
 }
 
 export default function DocsRoute() {
-  const { version, tanstackDocsConfig } = useLoaderData<typeof loader>()
-  let config = useRangerDocsConfig(tanstackDocsConfig)
+  const { version, config } = useLoaderData<typeof loader>()
+
   return (
     <DocsLayout
       name="Ranger"
@@ -43,6 +45,9 @@ export default function DocsRoute() {
       colorTo={colorTo}
       textColor={textColor}
       config={config}
+      frameworks={frameworks}
+      availableVersions={availableVersions}
+      localMenu={localMenu}
     >
       <Outlet />
     </DocsLayout>
