+++ app/utils/config.ts
@@ -1,9 +1,5 @@
-import { useMemo } from 'react'
-import { useMatches, useNavigate, useParams } from '@remix-run/react'
 import { z } from 'zod'
 import { fetchRepoFile } from './documents.server'
-import { generatePath } from '~/utils/utils'
-import type { AvailableOptions } from '~/components/Select'
 
 export type FrameworkMenu = {
   framework: string
@@ -73,118 +69,3 @@ export async function getTanstackDocsConfig(repo: string, branch: string) {
     throw new Error('Invalid docs/config.json file')
   }
 }
-
-/**
- * Use framework in URL path
- * Otherwise use framework in localStorage if it exists for this project
- * Otherwise fallback to react
- */
-export function useCurrentFramework(frameworks: AvailableOptions) {
-  const { framework: paramsFramework } = useParams()
-  const localStorageFramework = localStorage.getItem('framework')
-
-  return paramsFramework
-    ? paramsFramework
-    : localStorageFramework
-    ? frameworks[localStorageFramework]
-      ? localStorageFramework
-      : 'react'
-    : 'react'
-}
-
-export const useDocsConfig = ({
-  config,
-  frameworks,
-  availableVersions,
-  localMenu,
-}: {
-  config: ConfigSchema
-  frameworks: AvailableOptions
-  availableVersions: string[]
-  localMenu: MenuItem
-}) => {
-  const matches = useMatches()
-  const match = matches[matches.length - 1]
-  const params = useParams()
-  const version = params.version!
-  const framework = useCurrentFramework(frameworks)
-  const navigate = useNavigate()
-
-  const frameworkMenuItems =
-    config.frameworkMenus.find((d) => d.framework === framework)?.menuItems ??
-    []
-
-  const frameworkConfig = useMemo(() => {
-    return {
-      label: 'Framework',
-      selected: frameworks[framework] ? framework : 'react',
-      available: frameworks,
-      onSelect: (option: { label: string; value: string }) => {
-        const url = generatePath(match.id, {
-          ...match.params,
-          framework: option.value,
-        })
-
-        localStorage.setItem('framework', option.value)
-
-        navigate(url)
-      },
-    }
-  }, [frameworks, framework, match, navigate])
-
-  const versionConfig = useMemo(() => {
-    const available = availableVersions.reduce(
-      (acc: AvailableOptions, version) => {
-        acc[version] = {
-          label: version,
-          value: version,
-        }
-        return acc
-      },
-      {
-        latest: {
-          label: 'Latest',
-          value: 'latest',
-        },
-      }
-    )
-
-    return {
-      label: 'Version',
-      selected: version,
-      available,
-      onSelect: (option: { label: string; value: string }) => {
-        const url = generatePath(match.id, {
-          ...match.params,
-          version: option.value,
-        })
-        navigate(url)
-      },
-    }
-  }, [version, match, navigate, availableVersions])
-
-  return {
-    ...config,
-    menu: [
-      localMenu,
-      // Merge the two menus together based on their group labels
-      ...config.menu.map((d) => {
-        const match = frameworkMenuItems.find((d2) => d2.label === d.label)
-        return {
-          label: d.label,
-          children: [
-            ...d.children.map((d) => ({ ...d, badge: 'core' })),
-            ...(match?.children ?? []).map((d) => ({ ...d, badge: framework })),
-          ],
-        }
-      }),
-      ...frameworkMenuItems.filter(
-        (d) => !config.menu.find((dd) => dd.label === d.label)
-      ),
-    ].filter(Boolean),
-    frameworkConfig,
-    versionConfig,
-  }
-}
-
-export type DocsConfig = ReturnType<typeof useDocsConfig>
